+++ src/main/java/io/github/cottonmc/templates/model/RetexturingBakedModel.java
@@ -19,36 +19,37 @@ import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.BlockRenderView;
 
-import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.function.Supplier;
 
-public abstract class RetexturingBakedModel extends ForwardingBakedModel {
-	public RetexturingBakedModel(BakedModel baseModel, TemplateAppearanceManager tam, ModelBakeSettings settings, BlockState itemModelState, boolean ao) {
+public class RetexturingBakedModel extends ForwardingBakedModel {
+	public RetexturingBakedModel(BakedModel baseModel, Mesh baseMesh, TemplateAppearanceManager tam, ModelBakeSettings settings, boolean ao) {
 		this.wrapped = baseModel; //field from the superclass; vanilla getQuads etc will delegate through to this
-		
+		this.baseMesh = baseMesh;
 		this.tam = tam;
-		this.facePermutation = MeshTransformUtil.facePermutation(settings);
 		this.uvlock = settings.isUvLocked();
-		this.itemModelState = itemModelState;
 		this.ao = ao;
 	}
 	
 	protected final TemplateAppearanceManager tam;
-	protected final Map<Direction, Direction> facePermutation; //immutable
+	
+	//the "retexturable mesh", quads with tag 0 emit as-is, quads with tags 1-6 get retextured
+	protected final Mesh baseMesh;
+	
+	//pass MutableQuadView.BAKE_LOCK_UV as a bakeflag
 	protected final boolean uvlock;
-	protected final BlockState itemModelState;
+	
+	//use an ambient-occlusion enabled material
 	protected final boolean ao;
 	
-	protected record CacheKey(BlockState state, TemplateAppearance appearance) {}
-	protected final ConcurrentMap<CacheKey, Mesh> retexturedMeshes = new ConcurrentHashMap<>(); //mutable, append-only cache
+	//cache from template appearances ("blocks inside templates") to Mesh objects.
+	//never invalidated until you F3+T
+	protected final ConcurrentMap<TemplateAppearance, Mesh> retexturedMeshes = new ConcurrentHashMap<>();
 	
 	protected static final Direction[] DIRECTIONS = Direction.values();
 	protected static final Direction[] DIRECTIONS_AND_NULL = new Direction[DIRECTIONS.length + 1];
 	static { System.arraycopy(DIRECTIONS, 0, DIRECTIONS_AND_NULL, 0, DIRECTIONS.length); }
-	
-	protected abstract Mesh getBaseMesh(BlockState state);
 	
 	@Override
 	public boolean isVanillaAdapter() {
@@ -57,6 +58,8 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 	
 	@Override
 	public Sprite getParticleSprite() {
+		//We don't have access to the BlockRenderView at the moment, so return some junk.
+		//Mixins will override this for common particle types (breaking, sprinting, etc)
 		return tam.getDefaultAppearance().getSprite(Direction.UP);
 	}
 	
@@ -65,10 +68,9 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 		//FabricBlockView itf-inject
 		BlockState theme = blockView.getBlockEntityRenderData(pos) instanceof BlockState s ? s : null;
 		if(theme == null || theme.isAir()) {
-			getUntintedRetexturedMesh(new CacheKey(state, tam.getDefaultAppearance())).outputTo(context.getEmitter());
+			getUntintedRetexturedMesh(tam.getDefaultAppearance()).outputTo(context.getEmitter());
 			return;
 		} else if(theme.getBlock() == Blocks.BARRIER) {
-			//TODO i don't love putting this rare specialcase smack in the middle of the hot code path
 			return;
 		}
 		
@@ -77,12 +79,10 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 		theme = theme.getAppearance(blockView, pos, Direction.DOWN, theme, pos);
 		
 		TemplateAppearance ta = tam.getAppearance(theme);
+		Mesh untintedMesh = getUntintedRetexturedMesh(ta);
 		
+		//The specific tint might vary a lot (imagine grass color changing smoothly) so caching pre-tinted meshes is a waste of time.
 		int tint = 0xFF000000 | MinecraftClient.getInstance().getBlockColors().getColor(theme, blockView, pos, 0);
-		Mesh untintedMesh = getUntintedRetexturedMesh(new CacheKey(state, ta));
-		
-		//The specific tint might vary a lot; imagine grass color smoothly changing. Trying to bake the tint into
-		//the cached mesh will pollute it with a ton of single-use meshes with only slighly different colors.
 		if(tint == 0xFFFFFFFF) {
 			untintedMesh.outputTo(context.getEmitter());
 		} else {
@@ -107,7 +107,7 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 			tint = 0xFFFFFFFF;
 		}
 		
-		Mesh untintedMesh = getUntintedRetexturedMesh(new CacheKey(itemModelState, nbtAppearance));
+		Mesh untintedMesh = getUntintedRetexturedMesh(nbtAppearance);
 		
 		if(tint == 0xFFFFFFFF) {
 			untintedMesh.outputTo(context.getEmitter());
@@ -118,12 +118,12 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 		}
 	}
 	
-	protected Mesh getUntintedRetexturedMesh(CacheKey key) {
-		return retexturedMeshes.computeIfAbsent(key, this::createUntintedRetexturedMesh);
+	protected Mesh getUntintedRetexturedMesh(TemplateAppearance ta) {
+		return retexturedMeshes.computeIfAbsent(ta, this::createUntintedRetexturedMesh);
 	}
 	
-	protected Mesh createUntintedRetexturedMesh(CacheKey key) {
-		return MeshTransformUtil.pretransformMesh(getBaseMesh(key.state), new RetexturingTransformer(key.appearance));
+	protected Mesh createUntintedRetexturedMesh(TemplateAppearance ta) {
+		return MeshTransformUtil.pretransformMesh(baseMesh, new RetexturingTransformer(ta));
 	}
 	
 	protected class RetexturingTransformer implements RenderContext.QuadTransform {
@@ -140,15 +140,15 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 			int tag = quad.tag();
 			if(tag == 0) return true; //Pass the quad through unmodified.
 			
-			//The quad tag numbers were selected so this magic trick works:
-			Direction dir = facePermutation.get(DIRECTIONS[quad.tag() - 1]);
+			Direction dir = DIRECTIONS[quad.tag() - 1];
 			quad.spriteBake(ta.getSprite(dir), MutableQuadView.BAKE_NORMALIZED | ta.getBakeFlags(dir) | (uvlock ? MutableQuadView.BAKE_LOCK_UV : 0));
 			
 			return true;
 		}
 	}
 	
-	protected class TintingTransformer implements RenderContext.QuadTransform {
+	@SuppressWarnings("ClassCanBeRecord")
+	protected static class TintingTransformer implements RenderContext.QuadTransform {
 		protected TintingTransformer(TemplateAppearance ta, int tint) {
 			this.ta = ta;
 			this.tint = tint;
@@ -162,7 +162,7 @@ public abstract class RetexturingBakedModel extends ForwardingBakedModel {
 			int tag = quad.tag();
 			if(tag == 0) return true;
 			
-			Direction dir = facePermutation.get(DIRECTIONS[quad.tag() - 1]);
+			Direction dir = DIRECTIONS[quad.tag() - 1];
 			if(ta.hasColor(dir)) quad.color(tint, tint, tint, tint);
 			
 			return true;